Well friends, what have we here? An initial commit? It sure looks like quite a lot of code for a first commit. That is because this code comes from the Mono project's mono-upnp svn module. I have heavily modified both the code and the project structure and don't want to commit those changes to the main SVN repo right now. This is more of my experimental repo which lets me make more sweeping changes. This may or may not become the primary repo for Mono.Upnp - we'll see. Sorry for being so awful about commiting this work. Hopefully git will make things a little easier.
So the state of the code. The client and server stacks are both
out of order. I have done a major re-write of the serialization/
deserialization stuff, opting for a custom reflection-based XML
solution. That is mostly in place and this commit includes unit
tests.

I will be commiting unit tests as I develop from here on out.
I would really like to have as close to 100% coverage as is
possible. This will hopefully be the last major re-write of the
API and core library. My goal is to have steady, cumlative progress
for developing the UPnP stack going forward. We shall see.