Added XmlNamespaceAttribute which allows you to define namespace prefixes before they need to be used. Also added tests for it, but the attribute is unsupported, so the tests will fail.
Added tests for XmlEnumAttribute. There is no support for this
attribute yet, so the tests will fail.

Added XML namespaces to the control types and slightly re-jiggered
StateVariable. I will be adding serialization tests for the
control types, then I will be fixing everything so that it
actually serializes/deserializes properly. Made changes to
DummyDeserializer in prep for scpd deserialization tests.