OK, I started off adding a simple UPnP client test and found out that the server wasn't serving the actual device description XML (I wasn't actually testing for that in the server tests - just 200 OK). I tracked the problem to the fact that the XmlWriter wasn't flushing in the XmlSerializer.GetBytes method (we were just using GetString for all of the serialization tests). It then occurs to me that the XmlWriter is using UTF-16 and we need UTF-8. The default UTF-8 encoder includes a non-printable sentinal character at the start of the string. So, the changes are:
The XmlSerializer now uses a UTF-8 encoding which does not generate
the UTF-8 signal character.

XmlSerializer.GetBytes now flushes the XmlWriter so that we
actually get bytes (!).

XmlSerializer.GetString now uses GetBytes so that the XML
declaration has the correct encoding="utf-8" attribute.

We have a unit test for XmlSerializer.GetBytes.

The server unit tests check to make sure that the server is
actually serving XML.

Other little things here and there.