Fixed a serialization bug
We may try to deserialize into a property that is not writable and fail instead
of just ignoring it if that's the case. Also fixed a bug that will prevent
actions from working if the SoapEnvelope contains a Header element because it
can't figure out how to create a default instance of it.