build: Various fixes to get "make distcheck" to work
Generate all the binaries in a toplevel bin/ folder. As the
generated makefiles wipe out the build folder during distclean, having
it set to the project folder leads to bad things.

Use $(srcdir) to find the various .snk files, so that they are found
during an out-of-tree build, as "make distcheck" does it.

Go over all the various assembly references, and do not require a
specific version when it's not needed.

Let MonoDevelop refresh the generated code in
src/Mono.Upnp/Mono.Upnp.GtkClient/gtk-gui/.